* master:
  README: cosmetic changes to match XDA thread
  AK2: underp opposite slot case
  README: update for multi-slot zip instructions
  AK2: add support for multi-slot zips - extend `reset_ak` with `reset_ak keep` which will keep files to be available to patch the other slot - add optional slot_select=inactive shell variable to force inactive slot during slot detection
  AK2: add U-Boot RAMDisk uImage support
  README: update for Intel OSIP mboot, customdd, touchups
  AK2: add $OS$ header Android image format support - add built-in support for mboot for Intel OSIP devices (see: https://github.com/osm0sis/mboot)
  AK2: fix dump failures for unpackelf and dumpimage (with ramdisk)
  AK2: add AOSP boot_img_hdr_v2 support (READ NOTES) - AOSP has added --dtb, so to avoid conflicts *unpackbootimg and unpackelf have had what was previously dumped as *-dtb renamed to *-dt - the same corresponds to any replacement version of these included in the zip root, i.e. they are now dt or dt.img for the old style from before hdr_v1, and dtb or dtb.img for the new style for hdr_v2
  AK2: fix unpack for latest U-Boot tools (v2019.04) update - v2019.04 changed the order and arguments for the U-Boot dumpimage tool
  tools: update to latest builds, including busybox (1.30.1), lz4 (1.9.0) Compiled from my own updated forks: https://github.com/osm0sis/mkbootimg https://github.com/osm0sis/android-busybox-ndk
  AK2: fix customdd detection logic
  AK2: add backwards compat for crazy devices hacking their mmcblk0 see comments here: https://github.com/osm0sis/AnyKernel2/pull/38#issuecomment-479577751
  AK2: optimize write operations
  AK2: workaround for U-Boot dumpimage failing on untrimmed uImage dumps https://lists.denx.de/pipermail/u-boot/2019-January/355670.html
  AK2: fix ramdisk repack and mtk signing when uncompressed cpio
  README: document cpio compression support
  AK2: add support for uncompressed cpio ramdisks ramdisk_compression=cpio or =none to force
  Backend: support all half-sensible ROM version formats 9 9.0 9.0.0
  Backend: workaround bad ROM versioning practices in supported check - some ROMs have taken it upon themselves to go against AOSP, i.e. where starting with Pie AOSP now versions it as only 9, some ROMs are still making it 9.0 - introduce integer 2 float function that will add a decimal place to all for comparisons - this should support the majority of ROMs, but anything crazy like 9.0.0, or crazier like 09.0.0 should be frowned upon by the community and the ROM devs should be made to get it together and follow AOSP
  Backend: clean up mounting - only mount system when not already mounted (avoids harmless stderr during booted flashes) - optimize module targeting and mount logic further
  README: touchups to supported.versions documentation (Android 8.1 was 8.1.0 in build.prop ;) )
  Revert "anykernel: fix ramdisk/sbin permissions"
  README: document supported.versions
  anykernel: add supported.versions
  Backend: introduce Android version check - checks the supported.versions prop in anykernel.sh and matches against it to prevent users flashing incorrect kernels for ROM versions - supported.versions can be a list, e.g. "7.1.2, 8.0, 8.1" or a range, e.g. "8.1 - 9" of versions in the correct formatting for that Android version in ro.build.version.release in /system/build.prop
  README: add replace_string scope + note re: extra device.name# entries
  AK2: allow replace_string to optionally global scope replace
  AK2+Backend: fix file_getprop truncating if prop value contained = See comments: https://github.com/osm0sis/AnyKernel2/pull/35#issuecomment-447648900
  Backend: cleanups
  Backend: more robust getprop checks and replacements - getprop failures are not only a system-as-root device issue, as it now also happens on OnePlus 5T in recovery when ROM is Pie Open Beta - this appears to come down to a libc mismatch between recovery and ROM resulting in direct property checks failing but getprop without arguments printing all properties working - revert most of https://github.com/osm0sis/AnyKernel2/commit/b1524da099efa716616a8eb3357876bca6a73540 since this should remain resolved with the new implementation - keep the old check and file_getprop fallback for devices without getprop entirely - add new getprop fallback that parses the full getprop output
  README: document device.name<num> limit removal
  Backend: allow infinite device.name<num>= device check properties (#34)
  Backend: only attempt module bind remount if already mounted - blindly attempting to remount broke module placement on some ROMs (e.g. Pocophone MIUI stock)
  Backend: attempt remounting module location as well - some custom ROMs on some devices (e.g. Lineage on Pocophone) bind mount modules in /system over modules on /vendor, so remount rw at the bind target to ensure the bind source is also remounted and replaced - toybox mount doesn't accept the /. prefixed paths from find so ensure we use busybox for mounts

+++ tools/ak2-core.sh
@@ -1,4 +1,6 @@
 ## AnyKernel methods (DO NOT CHANGE)
+# osm0sis @ xda-developers
+
 # set up extracted files and directories
 ramdisk=/tmp/anykernel/ramdisk;
 bin=/tmp/anykernel/tools;
@@ -18,29 +20,42 @@ ui_print() { echo -e "ui_print $1\nui_print" > $OUTFD; }
 contains() { test "${1#*$2}" != "$1" && return 0 || return 1; }
 
 # file_getprop <file> <property>
-file_getprop() { grep "^$2=" "$1" | cut -d= -f2; }
+file_getprop() { grep "^$2=" "$1" | cut -d= -f2-; }
 
-# reset anykernel directory
+# reset_ak [keep]
 reset_ak() {
   local i;
-  rm -rf $(dirname /tmp/anykernel/*-files/current)/ramdisk;
-  for i in $ramdisk $split_img /tmp/anykernel/rdtmp /tmp/anykernel/boot.img /tmp/anykernel/*-new*; do
-    cp -af $i $(dirname /tmp/anykernel/*-files/current);
-  done;
-  rm -rf $ramdisk $split_img $patch /tmp/anykernel/rdtmp /tmp/anykernel/boot.img /tmp/anykernel/*-new* /tmp/anykernel/*-files/current;
+  if [ ! "$1" == "keep" ]; then
+    rm -rf $(dirname /tmp/anykernel/*-files/current)/ramdisk;
+    for i in $ramdisk $split_img /tmp/anykernel/rdtmp /tmp/anykernel/boot.img /tmp/anykernel/*-new*; do
+      cp -af $i $(dirname /tmp/anykernel/*-files/current);
+    done;
+  fi;
+  rm -rf $ramdisk $split_img /tmp/anykernel/boot.img /tmp/anykernel/*-new* /tmp/anykernel/*-files/current;
+  if [ "$1" == "keep" ]; then
+    mv -f /tmp/anykernel/rdtmp $ramdisk;
+  else
+    rm -rf $patch /tmp/anykernel/rdtmp;
+  fi;
   . /tmp/anykernel/tools/ak2-core.sh $FD;
 }
 
 # dump boot and extract ramdisk
 split_boot() {
-  local nooktest nookoff dumpfail;
+  local nooktest nookoff uimgsize dumpfail;
   if [ ! -e "$(echo $block | cut -d\  -f1)" ]; then
     ui_print " "; ui_print "Invalid partition. Aborting..."; exit 1;
+  fi;
+  if [ "$(echo $block | grep ' ')" ]; then
+    block=$(echo $block | cut -d\  -f1);
+    customdd=$(echo $block | cut -d\  -f2-);
+  elif [ ! "$customdd" ]; then
+    local customdd="bs=1048576";
   fi;
   if [ -f "$bin/nanddump" ]; then
     $bin/nanddump -f /tmp/anykernel/boot.img $block;
   else
-    dd if=$block of=/tmp/anykernel/boot.img;
+    dd if=$block of=/tmp/anykernel/boot.img $customdd;
   fi;
   nooktest=$(strings /tmp/anykernel/boot.img | grep -E 'Red Loader|Green Loader|Green Recovery|eMMC boot.img|eMMC recovery.img|BauwksBoot');
   if [ "$nooktest" ]; then
@@ -59,8 +74,20 @@ split_boot() {
       cp -f $split_img/elftool_out/header $split_img/boot.img-header;
     fi;
     $bin/unpackelf -i /tmp/anykernel/boot.img -o $split_img;
+    test $? != 0 && dumpfail=1;
     mv -f $split_img/boot.img-ramdisk.cpio.gz $split_img/boot.img-ramdisk.gz;
+  elif [ -f "$bin/mboot" ]; then
+    $bin/mboot -u -f /tmp/anykernel/boot.img -d $split_img;
+    test $? != 0 && dumpfail=1;
+    mv -f $split_img/cmdline.txt $split_img/boot.img-cmdline;
+    mv -f $split_img/kernel $split_img/boot.img-zImage;
+    mv -f $split_img/ramdisk.cpio.gz $split_img/boot.img-ramdisk.gz;
   elif [ -f "$bin/dumpimage" ]; then
+    uimgsize=$(($(printf '%d\n' 0x$(hexdump -n 4 -s 12 -e '16/1 "%02x""\n"' /tmp/anykernel/boot.img)) + 64));
+    if [ "$(wc -c < /tmp/anykernel/boot.img)" != "$uimgsize" ]; then
+      mv -f /tmp/anykernel/boot.img /tmp/anykernel/boot-orig.img;
+      dd bs=$uimgsize count=1 conv=notrunc if=/tmp/anykernel/boot-orig.img of=/tmp/anykernel/boot.img;
+    fi;
     $bin/dumpimage -l /tmp/anykernel/boot.img;
     $bin/dumpimage -l /tmp/anykernel/boot.img > $split_img/boot.img-header;
     grep "Name:" $split_img/boot.img-header | cut -c15- > $split_img/boot.img-name;
@@ -70,11 +97,12 @@ split_boot() {
     grep "Type:" $split_img/boot.img-header | cut -d\( -f2 | cut -d\) -f1 | cut -d\  -f1 | cut -d- -f1 > $split_img/boot.img-comp;
     grep "Address:" $split_img/boot.img-header | cut -c15- > $split_img/boot.img-addr;
     grep "Point:" $split_img/boot.img-header | cut -c15- > $split_img/boot.img-ep;
-    $bin/dumpimage -i /tmp/anykernel/boot.img -p 0 $split_img/boot.img-zImage;
+    $bin/dumpimage -p 0 -o $split_img/boot.img-zImage /tmp/anykernel/boot.img;
     test $? != 0 && dumpfail=1;
-    if [ "$(cat $split_img/boot.img-type)" == "Multi" ]; then
-      $bin/dumpimage -i /tmp/anykernel/boot.img -p 1 $split_img/boot.img-ramdisk.gz;
-    fi;
+    case $(cat $split_img/boot.img-type) in
+      Multi) $bin/dumpimage -p 1 -o $split_img/boot.img-ramdisk.gz /tmp/anykernel/boot.img;;
+      RAMDisk) mv -f $split_img/boot.img-zImage $split_img/boot.img-ramdisk.gz;;
+    esac;
     test $? != 0 && dumpfail=1;
   elif [ -f "$bin/rkcrc" ]; then
     dd bs=4096 skip=8 iflag=skip_bytes conv=notrunc if=/tmp/anykernel/boot.img of=$split_img/boot.img-ramdisk.gz;
@@ -86,13 +114,13 @@ split_boot() {
   if [ $? != 0 -o "$dumpfail" ]; then
     ui_print " "; ui_print "Dumping/splitting image failed. Aborting..."; exit 1;
   fi;
-  if [ -f "$bin/unpackelf" -a -f "$split_img/boot.img-dtb" ]; then
-    case $(od -ta -An -N4 $split_img/boot.img-dtb | sed -e 's/del //' -e 's/   //g') in
+  if [ -f "$bin/unpackelf" -a -f "$split_img/boot.img-dt" ]; then
+    case $(od -ta -An -N4 $split_img/boot.img-dt | sed -e 's/del //' -e 's/   //g') in
       QCDT|ELF) ;;
       *) gzip $split_img/boot.img-zImage;
          mv -f $split_img/boot.img-zImage.gz $split_img/boot.img-zImage;
-         cat $split_img/boot.img-dtb >> $split_img/boot.img-zImage;
-         rm -f $split_img/boot.img-dtb;;
+         cat $split_img/boot.img-dt >> $split_img/boot.img-zImage;
+         rm -f $split_img/boot.img-dt;;
     esac;
   fi;
 }
@@ -111,13 +139,19 @@ unpack_ramdisk() {
     '   B   Z   h'*) compext="bz2"; unpackcmd="bzip2";;
     ' stx   !   L can') compext="lz4-l"; unpackcmd="$bin/lz4";;
     ' etx   !   L can'|' eot   "   M can') compext="lz4"; unpackcmd="$bin/lz4";;
+    '   0   7   0   7') compext=""; unpackcmd="cat";;
+    '') ui_print " "; ui_print "Ramdisk not found in image. Aborting..."; exit 1;;
     *) ui_print " "; ui_print "Unknown ramdisk compression. Aborting..."; exit 1;;
   esac;
-  mv -f $split_img/boot.img-ramdisk.gz $split_img/boot.img-ramdisk.cpio.$compext;
+  if [ "$compext" ]; then
+    compext=.$compext;
+    unpackcmd="$unpackcmd -dc";
+  fi;
+  mv -f $split_img/boot.img-ramdisk.gz $split_img/boot.img-ramdisk.cpio$compext;
   mkdir -p $ramdisk;
   chmod 755 $ramdisk;
   cd $ramdisk;
-  $unpackcmd -dc $split_img/boot.img-ramdisk.cpio.$compext | EXTRACT_UNSAFE_SYMLINKS=1 cpio -i -d;
+  $unpackcmd $split_img/boot.img-ramdisk.cpio$compext | EXTRACT_UNSAFE_SYMLINKS=1 cpio -i -d;
   if [ $? != 0 -o -z "$(ls $ramdisk)" ]; then
     ui_print " "; ui_print "Unpacking ramdisk failed. Aborting..."; exit 1;
   fi;
@@ -132,7 +166,8 @@ dump_boot() {
 repack_ramdisk() {
   local compext repackcmd;
   case $ramdisk_compression in
-    auto|"") compext=`echo $split_img/*-ramdisk.cpio.* | rev | cut -d. -f1 | rev`;;
+    auto|"") compext=`echo $split_img/*-ramdisk.cpio* | rev | cut -d. -f1 | rev`; test "$compext" == "cpio" && compext="";;
+    none|cpio) compext="";;
     *) compext=$ramdisk_compression;;
   esac;
   case $compext in
@@ -143,20 +178,25 @@ repack_ramdisk() {
     bz2) repackcmd="bzip2";;
     lz4-l) repackcmd="$bin/lz4 -l";;
     lz4) repackcmd="$bin/lz4";;
+    "") repackcmd="cat";;
   esac;
+  if [ "$compext" ]; then
+    compext=.$compext;
+    repackcmd="$repackcmd -9c";
+  fi;
   if [ -f "$bin/mkbootfs" ]; then
-    $bin/mkbootfs $ramdisk | $repackcmd -9c > /tmp/anykernel/ramdisk-new.cpio.$compext;
+    $bin/mkbootfs $ramdisk | $repackcmd > /tmp/anykernel/ramdisk-new.cpio$compext;
   else
     cd $ramdisk;
-    find . | cpio -H newc -o | $repackcmd -9c > /tmp/anykernel/ramdisk-new.cpio.$compext;
+    find . | cpio -H newc -o | $repackcmd > /tmp/anykernel/ramdisk-new.cpio$compext;
   fi;
   if [ $? != 0 ]; then
     ui_print " "; ui_print "Repacking ramdisk failed. Aborting..."; exit 1;
   fi;
   cd /tmp/anykernel;
   if [ -f "$bin/mkmtkhdr" ]; then
-    $bin/mkmtkhdr --rootfs ramdisk-new.cpio.$compext;
-    mv -f ramdisk-new.cpio.$compext-mtk ramdisk-new.cpio.$compext;
+    $bin/mkmtkhdr --rootfs ramdisk-new.cpio$compext;
+    mv -f ramdisk-new.cpio$compext-mtk ramdisk-new.cpio$compext;
   fi;
 }
 flash_dtbo() {
@@ -175,13 +215,12 @@ flash_dtbo() {
       $bin/flash_erase $dtbo_block 0 0;
       $bin/nandwrite -p $dtbo_block /tmp/anykernel/$dtbo;
     else
-      dd if=/dev/zero of=$dtbo_block 2>/dev/null;
-      dd if=/tmp/anykernel/$dtbo of=$dtbo_block;
+      cat /tmp/anykernel/$dtbo /dev/zero > $dtbo_block 2>/dev/null;
     fi;
   fi;
 }
 flash_boot() {
-  local name arch os type comp addr ep cmdline cmd board base pagesize kerneloff ramdiskoff tagsoff osver oslvl second secondoff hash unknown i kernel rd dtb rpm pk8 cert avbtype dtbo dtbo_block;
+  local name arch os type comp addr ep cmdline cmd board base pagesize kerneloff ramdiskoff tagsoff dtboff osver oslvl hdrver second secondoff recoverydtbo hash unknown i kernel rd dtb dt rpm part0 part1 pk8 cert avbtype dtbo dtbo_block;
   cd $split_img;
   if [ -f "$bin/mkimage" ]; then
     name=`cat *-name`;
@@ -206,6 +245,10 @@ flash_boot() {
     ramdiskoff=`cat *-ramdiskoff`;
     if [ -f *-tagsoff ]; then
       tagsoff=`cat *-tagsoff`;
+    fi;
+    if [ -f *-dtboff ]; then
+      dtboff=`cat *-dtboff`;
+      dtboff="--dtb_offset $dtboff";
     fi;
     if [ -f *-osversion ]; then
       osver=`cat *-osversion`;
@@ -245,23 +288,33 @@ flash_boot() {
     kernel=`ls *-zImage`;
     kernel=$split_img/$kernel;
   fi;
-  if [ -f /tmp/anykernel/ramdisk-new.cpio.* ]; then
-    rd=`echo /tmp/anykernel/ramdisk-new.cpio.*`;
+  if [ -f /tmp/anykernel/ramdisk-new.cpio* ]; then
+    rd=`echo /tmp/anykernel/ramdisk-new.cpio*`;
   else
     rd=`ls *-ramdisk.*`;
     rd="$split_img/$rd";
   fi;
-  for i in dtb dt.img; do
+  for i in dtb dtb.img; do
     if [ -f /tmp/anykernel/$i ]; then
-      dtb="--dt /tmp/anykernel/$i";
-      rpm="/tmp/anykernel/$i,rpm";
+      dtb="--dtb /tmp/anykernel/$i";
       break;
     fi;
   done;
   if [ ! "$dtb" -a -f *-dtb ]; then
     dtb=`ls *-dtb`;
-    rpm="$split_img/$dtb,rpm";
-    dtb="--dt $split_img/$dtb";
+    dtb="--dtb $split_img/$dtb";
+  fi;
+  for i in dt dt.img; do
+    if [ -f /tmp/anykernel/$i ]; then
+      dt="--dt /tmp/anykernel/$i";
+      rpm="/tmp/anykernel/$i,rpm";
+      break;
+    fi;
+  done;
+  if [ ! "$dt" -a -f *-dt ]; then
+    dt=`ls *-dt`;
+    rpm="$split_img/$dt,rpm";
+    dt="--dt $split_img/$dt";
   fi;
   cd /tmp/anykernel;
   if [ -f "$bin/mkmtkhdr" ]; then
@@ -271,16 +324,25 @@ flash_boot() {
     esac;
   fi;
   if [ -f "$bin/mkimage" ]; then
-    test "$type" == "Multi" && uramdisk=":$rd";
-    $bin/mkimage -A $arch -O $os -T $type -C $comp -a $addr -e $ep -n "$name" -d $kernel$uramdisk boot-new.img;
+    part0=$kernel;
+    case $type in
+      Multi) part1=":$rd";;
+      RAMDisk) part0=$rd;;
+    esac;
+    $bin/mkimage -A $arch -O $os -T $type -C $comp -a $addr -e $ep -n "$name" -d $part0$part1 boot-new.img;
   elif [ -f "$bin/elftool" ]; then
     $bin/elftool pack -o boot-new.img header=$split_img/boot.img-header $kernel $rd,ramdisk $rpm $cmd;
+  elif [ -f "$bin/mboot" ]; then
+    cp -f $split_img/boot.img-cmdline $split_img/cmdline.txt;
+    cp -f $kernel $split_img/kernel;
+    cp -f $rd $split_img/ramdisk.cpio.gz;
+    $bin/mboot -d $split_img -f boot-new.img;
   elif [ -f "$bin/rkcrc" ]; then
     $bin/rkcrc -k $rd boot-new.img;
   elif [ -f "$bin/pxa-mkbootimg" ]; then
-    $bin/pxa-mkbootimg --kernel $kernel --ramdisk $rd $second --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset "$tagsoff" --unknown $unknown $dtb --output boot-new.img;
+    $bin/pxa-mkbootimg --kernel $kernel --ramdisk $rd $second --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset "$tagsoff" --unknown $unknown $dt --output boot-new.img;
   else
-    $bin/mkbootimg --kernel $kernel --ramdisk $rd $second $recoverydtbo --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset "$tagsoff" --os_version "$osver" --os_patch_level "$oslvl" --header_version "$hdrver" $hash $dtb --output boot-new.img;
+    $bin/mkbootimg --kernel $kernel --ramdisk $rd $second $dtb $recoverydtbo --cmdline "$cmdline" --board "$board" --base $base --pagesize $pagesize --kernel_offset $kerneloff --ramdisk_offset $ramdiskoff $secondoff --tags_offset "$tagsoff" $dtboff --os_version "$osver" --os_patch_level "$oslvl" --header_version "$hdrver" $hash $dt --output boot-new.img;
   fi;
   if [ $? != 0 ]; then
     ui_print " "; ui_print "Repacking image failed. Aborting..."; exit 1;
@@ -338,9 +400,11 @@ flash_boot() {
   if [ -f "$bin/flash_erase" -a -f "$bin/nandwrite" ]; then
     $bin/flash_erase $block 0 0;
     $bin/nandwrite -p $block /tmp/anykernel/boot-new.img;
+  elif [ "$customdd" ]; then
+    dd if=/dev/zero of=$block $customdd 2>/dev/null;
+    dd if=/tmp/anykernel/boot-new.img of=$block $customdd;
   else
-    dd if=/dev/zero of=$block 2>/dev/null;
-    dd if=/tmp/anykernel/boot-new.img of=$block;
+    cat /tmp/anykernel/boot-new.img /dev/zero > $block 2>/dev/null;
   fi;
 }
 write_boot() {
@@ -355,10 +419,11 @@ backup_file() { test ! -f $1~ && cp $1 $1~; }
 # restore_file <file>
 restore_file() { test -f $1~ && mv -f $1~ $1; }
 
-# replace_string <file> <if search string> <original string> <replacement string>
+# replace_string <file> <if search string> <original string> <replacement string> <scope>
 replace_string() {
+  test "$5" == "global" && local scope=g;
   if [ -z "$(grep "$2" $1)" ]; then
-    sed -i "s;${3};${4};" $1;
+    sed -i "s;${3};${4};${scope}" $1;
   fi;
 }
 
@@ -550,6 +615,16 @@ case $is_slot_device in
       slot=$(getprop ro.boot.slot 2>/dev/null);
       test ! "$slot" && slot=$(grep -o 'androidboot.slot=.*$' /proc/cmdline | cut -d\  -f1 | cut -d= -f2);
       test "$slot" && slot=_$slot;
+    fi;
+    if [ "$slot" ]; then
+      case $slot_select in
+        inactive)
+          case $slot in
+            _a) slot=_b;;
+            _b) slot=_a;;
+          esac;
+        ;;
+      esac;
     fi;
     if [ ! "$slot" -a "$is_slot_device" == 1 ]; then
       ui_print " "; ui_print "Unable to determine active boot slot. Aborting..."; exit 1;
